+++ sale_order_volume/models/__init__.py
@@ -1,2 +1,4 @@
+from . import product_category_volume
 from . import res_config_settings
+from . import sale_order_line
 from . import sale_order
