+++ sale_order_volume/tests/test_sale_order_volume.py
@@ -2,13 +2,10 @@
 #   Robin Keunen <robin@coopiteasy.be>
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
-import unittest
-
 from odoo.tests import TransactionCase
 
 
 class SaleOrderVolumeCase(TransactionCase):
-    @unittest.expectedFailure
     def test_sale_order_volumes(self):
         sale_order = self.browse_ref("sale.sale_order_4")
         product = self.browse_ref("product.product_delivery_01")
@@ -28,7 +25,7 @@ class SaleOrderVolumeCase(TransactionCase):
         )
 
         self.assertEqual(service_volume.volume, 0)
-        self.assertEqual(service_volume.pallet_count, 1)
+        self.assertEqual(service_volume.pallet_count, 0)
         self.assertEqual(office_furniture_volume.volume, 15.6)
         # (15.6 (volume) // 1.75 (pallet volume)) + 1 = 9
         self.assertEqual(office_furniture_volume.pallet_count, 9)
